# client-sdks/manager/rust/openapi-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alien Manager API",
    "description": "Control plane for Alien applications. Manages deployments, releases, commands, and telemetry.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/commands": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Create a new command",
        "operationId": "create_command",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Command created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/leases": {
      "post": {
        "tags": [
          "leases"
        ],
        "summary": "Acquire leases for polling deployments",
        "operationId": "acquire_leases",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Leases acquired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaseResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/leases/{lease_id}/release": {
      "post": {
        "tags": [
          "leases"
        ],
        "summary": "Release a lease",
        "operationId": "release_lease",
        "parameters": [
          {
            "name": "lease_id",
            "in": "path",
            "description": "Lease identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lease released successfully"
          },
          "404": {
            "description": "Lease not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/{command_id}": {
      "get": {
        "tags": [
          "commands"
        ],
        "summary": "Get command status",
        "operationId": "get_command_status",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Command status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/{command_id}/payload": {
      "get": {
        "tags": [
          "commands"
        ],
        "summary": "Get command payload (params and response) from KV",
        "description": "Returns the raw params and response data stored in the manager's KV store.\nReturns 404 if neither params nor response exist for this command.",
        "operationId": "get_command_payload",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Command payload data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandPayloadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command payload not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "commands"
        ],
        "summary": "Store command payload data (params and/or response) directly into KV.",
        "description": "Bypasses the command registry — useful for populating demo data or\nmigrating payload data. Does not validate command existence or state.",
        "operationId": "store_command_payload",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorePayloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payload stored successfully"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/{command_id}/response": {
      "put": {
        "tags": [
          "deployments"
        ],
        "summary": "Submit response from deployment",
        "operationId": "submit_response",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response submitted successfully"
          },
          "400": {
            "description": "Invalid command or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/commands/{command_id}/upload-complete": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Mark upload as complete",
        "operationId": "upload_complete",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "description": "Command identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Upload marked complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCompleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid command or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/deployments/{id}": {
      "get": {
        "tags": [
          "deployments"
        ],
        "operationId": "get_deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deployment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "deployments"
        ],
        "operationId": "delete_deployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deployment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "Force delete without running cleanup (immediately removes record)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deployment deletion enqueued"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/v1/sync/acquire": {
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "acquire",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquireRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Deployments acquired for reconciliation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquireResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/sync/release": {
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "release",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Deployment lock released"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcquireResponse": {
        "type": "object",
        "required": [
          "deployments"
        ],
        "properties": {
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcquiredDeploymentResponse"
            }
          }
        }
      },
...
      "CommandPayloadResponse": {
        "type": "object",
        "description": "Payload response containing params and response data from KV",
        "required": [
          "commandId"
        ],
        "properties": {
          "commandId": {
            "type": "string"
          },
          "params": {
            "$ref": "#/components/schemas/BodySpec",
            "nullable": true
          },
          "response": {
            "$ref": "#/components/schemas/CommandResponse",
            "nullable": true
          }
        }
      },
...
      "CommandStatusResponse": {
        "type": "object",
        "description": "Response to status queries",
        "required": [
          "commandId",
          "state",
          "attempt"
        ],
        "properties": {
          "attempt": {
            "type": "integer",
            "format": "int32",
            "description": "Current attempt number",
            "minimum": 0
          },
          "commandId": {
            "type": "string",
            "description": "Command identifier"
          },
          "response": {
            "$ref": "#/components/schemas/CommandResponse",
            "description": "Response data (only for succeeded/failed state)",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/CommandState",
            "description": "Current command state"
          }
        }
      },
...
      "CreateCommandRequest": {
        "type": "object",
        "description": "Request to create a new command",
        "required": [
          "deploymentId",
          "command",
          "params"
        ],
        "properties": {
          "command": {
            "type": "string",
            "description": "Command name (e.g., \"generate-report\", \"sync-data\")"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "description": "Optional deadline for command completion",
            "nullable": true
          },
          "deploymentId": {
            "type": "string",
            "description": "Target deployment identifier"
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Optional idempotency key",
            "nullable": true
          },
          "params": {
            "$ref": "#/components/schemas/BodySpec",
            "description": "Command parameters (JSON, can be large)"
          }
        }
      },
      "CreateCommandResponse": {
        "type": "object",
        "description": "Response to command creation",
        "required": [
          "commandId",
          "state",
          "inlineAllowedUpTo",
          "next"
        ],
        "properties": {
          "commandId": {
            "type": "string",
            "description": "Unique command identifier"
          },
          "inlineAllowedUpTo": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum inline body size allowed",
            "minimum": 0
          },
          "next": {
            "type": "string",
            "description": "Next action for client: \"upload\" | \"poll\""
          },
          "state": {
            "$ref": "#/components/schemas/CommandState",
            "description": "Current command state"
          },
          "storageUpload": {
            "$ref": "#/components/schemas/StorageUpload",
            "description": "Storage upload info (only for storage mode)",
            "nullable": true
          }
        }
      },
...
      "DeploymentResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "platform",
          "status",
          "deploymentGroupId",
          "retryRequested",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "currentReleaseId": {
            "type": "string",
            "nullable": true
          },
          "deploymentGroup": {
            "$ref": "#/components/schemas/DeploymentGroupMinimal",
            "nullable": true
          },
          "deploymentGroupId": {
            "type": "string"
          },
          "desiredReleaseId": {
            "type": "string",
            "nullable": true
          },
          "environmentInfo": {},
          "error": {},
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "retryRequested": {
            "type": "boolean"
          },
          "runtimeMetadata": {},
          "stackSettings": {},
          "stackState": {},
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response wrapper for API endpoints",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LeaseRequest": {
        "type": "object",
        "description": "Request for acquiring leases",
        "required": [
          "deploymentId"
        ],
        "properties": {
          "deploymentId": {
            "type": "string",
            "description": "Deployment identifier"
          },
          "leaseSeconds": {
            "type": "integer",
            "format": "int64",
            "description": "Lease duration in seconds",
            "minimum": 0
          },
          "maxLeases": {
            "type": "integer",
            "description": "Maximum number of leases to acquire",
            "minimum": 0
          }
        }
      },
      "LeaseResponse": {
        "type": "object",
        "description": "Response to lease acquisition",
        "required": [
          "leases"
        ],
        "properties": {
          "leases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaseInfo"
            },
            "description": "Acquired leases (empty array if none available)"
          }
        }
      },
...
      "StorePayloadRequest": {
        "type": "object",
        "description": "Request to store payload data directly in KV by command_id.\n\nThis bypasses the normal command lifecycle (create → dispatch → respond)\nand writes params/response directly into KV. Used by the demo service\nto populate payload data for commands created outside the command flow.",
        "properties": {
          "params": {
            "$ref": "#/components/schemas/BodySpec",
            "nullable": true
          },
          "response": {
            "$ref": "#/components/schemas/CommandResponse",
            "nullable": true
          }
        }
      },
      "SubmitResponseRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommandResponse",
            "description": "The command response"
          }
        ],
        "description": "Request to submit a command response (from deployment)"
      },
...
      "UploadCompleteRequest": {
        "type": "object",
        "description": "Request to mark upload as complete",
        "required": [
          "size"
        ],
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Size of uploaded data",
            "minimum": 0
          }
        }
      },
      "UploadCompleteResponse": {
        "type": "object",
        "description": "Response to upload completion",
        "required": [
          "commandId",
          "state"
        ],
        "properties": {
          "commandId": {
            "type": "string",
            "description": "Command identifier"
          },
          "state": {
            "$ref": "#/components/schemas/CommandState",
            "description": "Updated command state"
          }
        }
      },
...
    }
  },
...
}